Fold extractor option parsing into greedy DAG module
