+++ src/lib.rs
@@ -45,7 +45,6 @@ pub mod scheduling;
 pub use scheduling::*;
 mod fresh_macro;
 
-mod extractor_option;
 mod greedy_dag_extract;
 mod secondary_map;
 pub use greedy_dag_extract::DagCostModel;
